修正

+++ AegisEngine/Time.cpp
@@ -7,4 +7,4 @@ DWORD CLOCK_TIMER::Time = 0;
 LARGE_INTEGER TIMER::delta_start;
 LARGE_INTEGER TIMER::delta_end;
 
-DWORD TIMER::time = 0;
+double TIMER::time = 0;
